+++ sentry/src/transports/reqwest.rs
@@ -6,7 +6,7 @@ use sentry_core::TransportOptions;
 
 use super::{
     tokio_thread::{TransportThread, TransportThreadOptions},
-    RateLimiter, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE,
+    RateLimiter, DEFAULT_CHANNEL_CAPACITY, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE,
 };
 
 use crate::{sentry_debug, ClientOptions, Envelope, Transport};
@@ -84,6 +84,7 @@ impl ReqwestHttpTransport {
                     https_proxy,
                     accept_invalid_certs,
                     client_report_recorder,
+                    transport_channel_capacity,
                     ..
                 },
             client,
@@ -192,6 +193,7 @@ impl ReqwestHttpTransport {
 
         let thread = TransportThreadOptions::new(send_fn)
             .with_client_report_recorder(client_report_recorder)
+            .with_capacity(transport_channel_capacity.unwrap_or(DEFAULT_CHANNEL_CAPACITY))
             .spawn_thread();
         Self { thread }
     }